backend/requirements.txt
# UrbanNest AI — Backend Requirements
# Python 3.11+

# Core API
fastapi==0.110.0
uvicorn[standard]==0.28.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.1

# Database
sqlalchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.5

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP clients
httpx==0.27.0
aiohttp==3.9.3

# Web scraping
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.42.0

# Task queue
celery[redis]==5.3.6
redis==5.0.2

# AI / ML
google-generativeai==0.5.0
langchain==0.1.13
langchain-google-genai==1.0.0
faiss-cpu==1.8.0
sentence-transformers==2.6.1
numpy==1.26.4
scikit-learn==1.4.1

# Utilities
python-slugify==8.0.4
Pillow==10.2.0
